Update comparison torchtune vs PEFT QLoRA
Finally figured out what the difference was. It had nothing to do with
bnb vs torchao or anything like that. It was simply that the actiation
checkpointing code in the torchtune script did nothing for transformers
models, which means that torchtune had activation checkpointing and PEFT
didn't. This explains why torchtune was so much more memory efficient.

Since it was not quite trivial to add activation checkpointing to
PEFT/transformers (this is usually handled by Trainer), I just disabled
it for torchtune to make the results comparible.

Now the memory advantage for torchtune is almost entirely gone. Using
transformers is still quite a lot faster.